code/requirements.txt
pip
Flask
cassandra-driver  
passlib
flask-login
requests
